Refactor code to internal folder and own folders
