+++ cmd/main.go
@@ -1,72 +1,10 @@
 package main
 
 import (
-	"net/http"
-	"os"
-	"time"
-
-	"github.com/ggicci/httpin"
+	"github.com/SongStitch/song-stitch/internal/server"
 	"github.com/joho/godotenv"
-	"github.com/justinas/alice"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/hlog"
 	"github.com/rs/zerolog/log"
 )
-
-func runServer() {
-	log := zerolog.New(os.Stdout).With().Timestamp().Logger()
-	c := alice.New()
-	c = c.Append(hlog.NewHandler(log))
-	c = c.Append(hlog.AccessHandler(func(r *http.Request, status, size int, duration time.Duration) {
-		hlog.FromRequest(r).Info().
-			Str("method", r.Method).
-			Stringer("url", r.URL).
-			Int("status", status).
-			Int("size", size).
-			Dur("duration", duration).
-			Msg("")
-	}))
-	c = c.Append(hlog.RemoteAddrHandler("ip"))
-	c = c.Append(hlog.UserAgentHandler("user_agent"))
-	c = c.Append(hlog.RefererHandler("referer"))
-	c = c.Append(hlog.RequestIDHandler("req_id", "Request-Id"))
-	h := c.
-		Append(httpin.NewInput(CollageRequest{})).
-		ThenFunc(collage)
-
-	router := http.NewServeMux()
-	router.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		http.ServeFile(w, r, "public/index.html")
-	})
-	router.Handle("/collage", h)
-
-	// serve files from public folder
-	fs := http.FileServer(http.Dir("public"))
-	router.Handle("/public/", http.StripPrefix("/public/", fs))
-
-	// serve robots.txt file
-	router.HandleFunc("/robots.txt", func(w http.ResponseWriter, r *http.Request) {
-		http.ServeFile(w, r, "public/robots.txt")
-	})
-
-	// serve humans.txt file
-	router.HandleFunc("/humans.txt", func(w http.ResponseWriter, r *http.Request) {
-		http.ServeFile(w, r, "public/humans.txt")
-	})
-
-	// nextRequestID := func() string {
-	// 	return fmt.Sprintf("%d", time.Now().UnixNano())
-	// }
-
-	server := &http.Server{
-		Addr:    ":8080",
-		Handler: router,
-	}
-	log.Info().Msg("Starting server...")
-	if err := server.ListenAndServe(); err != nil {
-		log.Fatal().Err(err).Msg("Startup failed")
-	}
-}
 
 func main() {
 	err := godotenv.Load()
@@ -74,5 +12,5 @@ func main() {
 		log.Info().Msg("Error loading .env file")
 	}
 
-	runServer()
+	server.RunServer()
 }
