+++ internal/api/api.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"context"
@@ -10,71 +10,13 @@ import (
 	"github.com/ggicci/httpin"
 	"github.com/go-playground/validator/v10"
 	"github.com/rs/zerolog"
+
+	"github.com/SongStitch/song-stitch/internal/collages"
+	"github.com/SongStitch/song-stitch/internal/generator"
+	"github.com/SongStitch/song-stitch/internal/session"
 )
-
-type Period string
-
-const (
-	OVERALL       Period = "overall"
-	SEVEN_DAYS    Period = "7day"
-	ONE_MONTH     Period = "1month"
-	THREE_MONTHS  Period = "3month"
-	SIX_MONTHS    Period = "6month"
-	TWELVE_MONTHS Period = "12month"
-)
-
-type CollageType string
-
-const (
-	ALBUM  CollageType = "album"
-	ARTIST CollageType = "artist"
-	TRACK  CollageType = "track"
-)
-
-var ErrTooManyImages = errors.New("too many images requested")
 
 // Seems stupid but was generated by copilot so ¯\_(ツ)_/¯
-func getPeriodFromStr(s string) Period {
-	switch s {
-	case "overall":
-		return OVERALL
-	case "7day":
-		return SEVEN_DAYS
-	case "1month":
-		return ONE_MONTH
-	case "3month":
-		return THREE_MONTHS
-	case "6month":
-		return SIX_MONTHS
-	case "12month":
-		return TWELVE_MONTHS
-	default:
-		return OVERALL
-	}
-}
-
-func validatePeriod(fl validator.FieldLevel) bool {
-	period := Period(fl.Field().String())
-	switch period {
-	case OVERALL, SEVEN_DAYS, ONE_MONTH, THREE_MONTHS, SIX_MONTHS, TWELVE_MONTHS:
-		return true
-	default:
-		return false
-	}
-}
-
-func getCollageTypeFromStr(s string) CollageType {
-	switch s {
-	case "album":
-		return ALBUM
-	case "artist":
-		return ARTIST
-	case "track":
-		return TRACK
-	default:
-		return ALBUM
-	}
-}
 
 type CollageRequest struct {
 	Rows          int    `in:"query=rows;default=3" validate:"required,gte=1,lte=15"`
@@ -92,44 +34,6 @@ type CollageRequest struct {
 	FontSize      int    `in:"query=fontsize;default=12" validate:"gte=8,lte=30"`
 }
 
-func generateCollageForAlbum(ctx context.Context, username string, period Period, count int, imageSize string, displayOptions DisplayOptions) (image.Image, error) {
-	albums, err := getAlbums(ctx, username, period, count, imageSize)
-	if err != nil {
-		return nil, err
-	}
-
-	downloadImages(ctx, albums)
-
-	return createCollage(ctx, albums, displayOptions)
-}
-
-func generateCollageForArtist(ctx context.Context, username string, period Period, count int, imageSize string, displayOptions DisplayOptions) (image.Image, error) {
-	if count > 100 {
-		return nil, ErrTooManyImages
-	}
-	artists, err := getArtists(ctx, username, period, count, imageSize)
-	if err != nil {
-		return nil, err
-	}
-
-	downloadImages(ctx, artists)
-
-	return createCollage(ctx, artists, displayOptions)
-}
-
-func generateCollageForTrack(ctx context.Context, username string, period Period, count int, imageSize string, displayOptions DisplayOptions) (image.Image, error) {
-	if count > 25 {
-		return nil, ErrTooManyImages
-	}
-	tracks, err := getTracks(ctx, username, period, count, imageSize)
-	if err != nil {
-		return nil, err
-	}
-
-	downloadImages(ctx, tracks)
-
-	return createCollage(ctx, tracks, displayOptions)
-}
 func generateCollage(ctx context.Context, request *CollageRequest, logger *zerolog.Logger) (image.Image, error) {
 	count := request.Rows * request.Columns
 	imageSize := "extralarge"
@@ -145,7 +49,7 @@ func generateCollage(ctx context.Context, request *CollageRequest, logger *zerol
 		imageDimension = 3
 	}
 
-	displayOptions := DisplayOptions{
+	displayOptions := generator.DisplayOptions{
 		ArtistName:     request.DisplayArtist,
 		AlbumName:      request.DisplayAlbum,
 		TrackName:      request.DisplayTrack,
@@ -160,21 +64,21 @@ func generateCollage(ctx context.Context, request *CollageRequest, logger *zerol
 		Columns:        request.Columns,
 	}
 
-	period := getPeriodFromStr(request.Period)
-	method := getCollageTypeFromStr(request.Method)
+	period := session.GetPeriodFromStr(request.Period)
+	method := session.GetCollageTypeFromStr(request.Method)
 	switch method {
-	case ALBUM:
-		return generateCollageForAlbum(ctx, request.Username, period, count, imageSize, displayOptions)
-	case ARTIST:
-		return generateCollageForArtist(ctx, request.Username, period, count, imageSize, displayOptions)
-	case TRACK:
-		return generateCollageForTrack(ctx, request.Username, period, count, imageSize, displayOptions)
+	case session.ALBUM:
+		return collages.GenerateCollageForAlbum(ctx, request.Username, period, count, imageSize, displayOptions)
+	case session.ARTIST:
+		return collages.GenerateCollageForArtist(ctx, request.Username, period, count, imageSize, displayOptions)
+	case session.TRACK:
+		return collages.GenerateCollageForTrack(ctx, request.Username, period, count, imageSize, displayOptions)
 	default:
 		return nil, errors.New("invalid collage type")
 	}
 }
 
-func collage(w http.ResponseWriter, r *http.Request) {
+func Collage(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	request := ctx.Value(httpin.Input).(*CollageRequest)
 	logger := zerolog.Ctx(ctx)
@@ -194,9 +98,9 @@ func collage(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		logger.Error().Err(err).Msg("Error occurred generating collage")
 		switch {
-		case err == ErrUserNotFound:
+		case err == session.ErrUserNotFound:
 			http.Error(w, "User not found", http.StatusNotFound)
-		case err == ErrTooManyImages:
+		case err == session.ErrTooManyImages:
 			http.Error(w, "Requested collage size is too large for the collage type", http.StatusBadRequest)
 		default:
 			http.Error(w, err.Error(), http.StatusInternalServerError)
