+++ internal/clients/interfaces.go
@@ -1,0 +1,7 @@
+package clients
+
+type LastFMResponse interface {
+	Append(l LastFMResponse) error
+	GetTotalPages() int
+	GetTotalFetched() int
+}
